#329 make external simplify function
